Move rocketchat settings to specific package
