+++ .meteor/packages
@@ -201,3 +201,4 @@ rocketchat:mailmessages
 juliancwirko:postcss
 littledata:synced-cron
 rocketchat:utils
+rocketchat:settings
