+++ packages/rocketchat-lib/package.js
@@ -32,6 +32,7 @@ Package.onUse(function(api) {
 	api.use('rocketchat:version');
 	api.use('rocketchat:logger');
 	api.use('rocketchat:mailer');
+	api.use('rocketchat:settings');
 	api.use('mizzao:timesync');
 	api.use('rocketchat:custom-oauth');
 	api.use('konecty:multiple-instances-status');
