+++ packages/rocketchat-settings/client/index.js
@@ -1,0 +1,5 @@
+import { settings } from '../lib/settings';
+
+export {
+	settings,
+};
