+++ packages/rocketchat-settings/package.js
@@ -1,0 +1,14 @@
+Package.describe({
+	name: 'rocketchat:settings',
+	version: '0.0.1',
+	summary: '',
+	git: '',
+});
+
+Package.onUse(function(api) {
+	api.use([
+		'ecmascript',
+	]);
+	api.mainModule('client/index.js', 'client');
+	api.mainModule('server/index.js', 'server');
+});
